Project 4: Build a data story Part 1
In this mid-term project, you’ll take the role of a data science consultant for a large production company and use the Movies Metadata dataset. The goal will be to apply the skills in structuring a data story and to practice defining your own problem statement and the analyses you plan to do. The goal of this project is to define your own problem statement and the flow of a powerful data story. You’ll create a ghost deck with the problem statement, potential analyses, and a set of potential recommendations.