Project 5: Build a data Story Part 2
In this project, you’ll continue to build on the mid-term project. Building on the ghost deck, you will perform various analyses to understand and find insights from the Movies Metadata. You will identify the limitations and biases in data that affect the recommendations. You will put it all together by building on your ghost deck, conducting the actual analysis, and creating a final presentation with final recommendations and an executive summary of key next steps.