Project 7: Animate a data story Part 2
In this project, you will build on your mid-term project from the World Bank Indicators data file to create an animated data story. In this final part of the project, you will create an animated data story and add an audio track to create a narrated finished product that you can add to your portfolio.